+++ frontend/src/App.vue
@@ -10,17 +10,7 @@
             </button>
         </div>
         <div class="map-area">
-            <div class="searchbox">
-                <div>
-                    <input type="text" value="포토 스팟" @keyup.enter="searchPlace">
-                </div>
-                <div class="results">
-                    <div class="place" v-for="rs in search.results" :key="rs.id" @click="showPlace(rs)">
-                        <h4>{{ rs.place_name }}</h4>
-                        <div class="addr">{{ rs.address_name }}</div>
-                    </div>
-                </div>
-            </div>
+            <SlideMenu :map-option="mapOption" @update-map-center="updateMapCenter" />
             <KakaoMap ref="kmap" class="kmap" :options="mapOption">
                 <template v-slot:overlay>
                     <div class="overlay-popup" ref="harborOverlay">
@@ -42,10 +32,12 @@ import KakaoMap from './components/map/KakaoMap.vue';
 import api from './service/api';
 import MarkerHandler from './components/map/marker-handler.js';
 import KakoOverlay from './components/map/overlay';
+import SlideMenu from './components/slideMenu/SlideMenu.vue';
 
 export default {
     components: {
         KakaoMap,
+        SlideMenu,
     },
     data() {
         return {
@@ -55,11 +47,6 @@ export default {
                     lng: 126.92761685591375,
                 },
                 level: 4,
-            },
-            search: {
-                keyword: null,
-                pgn: null,
-                results: [],
             },
             harbors: [],
             markers: null, // marker handler
@@ -94,19 +81,6 @@ export default {
         });
     },
     methods: {
-        searchPlace(e) {
-            const keyword = e.target.value.trim();
-            if(keyword.length === 0) {
-                return;
-            }
-
-            const ps = new window.kakao.maps.services.Places();
-            ps.keywordSearch(keyword, (data, status, pgn) => {
-                this.search.keyword = keyword;
-                this.search.pgn = pgn;
-                this.search.results = data;
-            });
-        },
         zoom(delta) {
             const level = Math.max(1, this.mapOption.level + delta);
             this.mapOption.level = level;
@@ -118,17 +92,13 @@ export default {
                 lng: harbor.lng,
             };
         },
-        showPlace(place) {
-            console.log(place);
-            
-            this.mapOption.center = {
-                lat: place.y,
-                lng: place.x 
-            };
-        },
         closeOverlay() {
             this.overlay.hide();
-        }
+        },
+        updateMapCenter(center) {
+            // mapOption의 center를 업데이트합니다
+            this.mapOption.center = center;
+        },
     },
 }
 </script>
@@ -174,7 +144,7 @@ button {
             }
         }
     }
-    .searchbox {
+    /* .searchbox {
         position: absolute;
         top: 0;
         left: 0;
@@ -196,7 +166,7 @@ button {
                 }
             }
         }
-    }
+    } */
     .kmap {
         flex: 1 1 auto;
         .overlay-popup {
